Update tinynav.js
Better way of setting active item back to selectbox, because working with indexes doesn't work for every situation.
Also added headerLink option, because the header can be selected back too. Default set to '/' which means the root of the site.
Added "activeSelected" setting to make active one selected again. Not based on the header text, that's weird!